+++ jsr.json
@@ -7,12 +7,20 @@
   "exports": {
     ".": "./index.ts",
     "./browser": "./browser.ts",
+    "./storage": "./storage.ts",
+    "./advanced": "./advanced.ts",
+    "./diagnostics": "./diagnostics.ts",
+    "./legacy": "./legacy.ts",
     "./sha1sync": "./src/infrastructure/adapters/sha1sync.ts"
   },
   "publish": {
     "include": [
       "index.ts",
       "browser.ts",
+      "storage.ts",
+      "advanced.ts",
+      "diagnostics.ts",
+      "legacy.ts",
       "rootCompatibility.ts",
       "src/**/*.ts",
       "src/**/*.d.ts",
