+++ test/conformance/v18CoordinateOpticPublicPath.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { openWarpWorldline } from '../../index.ts';
+import { openWarpWorldline } from '../../legacy.ts';
 import InMemoryGraphAdapter from '../../src/infrastructure/adapters/InMemoryGraphAdapter.ts';
 import { openRuntimeHostProduct } from '../../src/domain/warp/RuntimeHostProduct.ts';
 import type { PatchBuilder } from '../../src/domain/services/PatchBuilder.ts';
