+++ test/unit/domain/WarpCore.emit.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'vitest';
 import {
   WarpCore,
   InMemoryGraphAdapter,
-} from '../../../index.ts';
+} from '../../../legacy.ts';
 import { EFFECT_NODE_PREFIX } from '../../../src/domain/services/KeyCodec.ts';
 
 type WarpCoreWired = Awaited<ReturnType<typeof WarpCore.open>>;
