+++ test/unit/domain/WarpGraph.public-sync.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { openWarpGraph } from '../../../index.ts';
+import { openWarpGraph } from '../../../legacy.ts';
 import { createGitRepo } from '../../helpers/warpGraphTestUtils.ts';
 
 describe('WarpGraph public sync seam', { timeout: 20000 }, () => {
