+++ test/unit/domain/WarpWorldline.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { openWarpWorldline } from '../../../index.ts';
+import { openWarpWorldline } from '../../../legacy.ts';
 import InMemoryGraphAdapter from '../../../src/infrastructure/adapters/InMemoryGraphAdapter.ts';
 import WarpWorldline, { type WarpWorldlinePatchBuild } from '../../../src/domain/WarpWorldline.ts';
 import ProjectionHandle from '../../../src/domain/services/ProjectionHandle.ts';
