+++ test/unit/domain/publicReadApiExecutablePaths.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { openWarpWorldline } from '../../../index.ts';
+import { openWarpWorldline } from '../../../legacy.ts';
 import InMemoryGraphAdapter from '../../../src/infrastructure/adapters/InMemoryGraphAdapter.ts';
 import { openRuntimeHostProduct } from '../../../src/domain/warp/RuntimeHostProduct.ts';
