+++ test/unit/domain/services/query/BoundedSupportRule.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 import {
   BoundedSupportRule,
   QueryBuilder,
-} from '../../../../../index.ts';
+} from '../../../../../legacy.ts';
 import QueryError from '../../../../../src/domain/errors/QueryError.ts';
 import type {
   QueryNeighborEntry,
