+++ tsconfig.publish.json
@@ -11,6 +11,10 @@
   "include": [
     "index.ts",
     "browser.ts",
+    "storage.ts",
+    "advanced.ts",
+    "diagnostics.ts",
+    "legacy.ts",
     "src/**/*.ts",
     "src/globals.d.ts",
     "bin/**/*.ts",
