+++ tsconfig.src.json
@@ -3,6 +3,10 @@
   "include": [
     "index.ts",
     "browser.ts",
+    "storage.ts",
+    "advanced.ts",
+    "diagnostics.ts",
+    "legacy.ts",
     "src/**/*.ts",
     "src/globals.d.ts",
     "bin/warp-graph.ts",
